requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==22.0.0
packaging==24.1
platformdirs==4.2.2
PyJWT==2.8.0
pytz==2024.1
ratelimit==2.2.1
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2
